Move common functionality to base class
