+++ src/fontra/views/fontinfo/panel-sources.js
@@ -4,4 +4,10 @@ import { BaseInfoPanel } from "./panel-base.js";
 export class SourcesPanel extends BaseInfoPanel {
   static title = "Sources";
   static id = "sources-panel";
+
+  setupUI() {
+    this.panelElement.appendChild(
+      html.div({}, [`⚠️ under construction: placeholder for ${this.constructor.id}`])
+    );
+  }
 }
